6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,20 +1,19 @@
 #include <stdio.h>
 /**
- * main - Entry point 
- * Return: Always 0 (success)
+ * main - Entry point
+ * Return: Always 0 (Success)
  */
 int main(void)
 {
-	        char a;
-		int b;
-		long int c;
-		long d;
-		float e;
-printf("Size of a char: %i byte(s)\n",sizeof(a));
-printf("Size of a int: %i byte(s)\n",sizeof(b));
-printf("Size of a long int: %i byte(s)\n",sizeof(c));
-printf("Size of a long long int: %i byte(s)\n",sizeof(d));
-printf("Size of a float: %i byte(s)\n",sizeof(e));
+char a;
+int b;
+long int c;
+long long d;
+float e;
+printf("Size of a char: %d byte(s)\n", sizeof(a));
+printf("Size of a int: %d byte(s)\n", sizeof(b));
+printf("Size of a long int: %d byte(s)\n", sizeof(c));
+printf("Size of a long long int: %d byte(s)\n", sizeof(d));
+printf("Size of a float: %d byte(s)\n", sizeof(e));
 return (0);
 }
-
